move(definition): into /feature
