+++ server/src/embeddedLanguages/languageModes.ts
@@ -13,7 +13,7 @@ import {
   TextDocument,
 } from "vscode-html-languageservice";
 import { TextDocumentPositionParams } from "vscode-languageserver";
-import { DefinitionResult } from "../definition/getDefinition";
+import { DefinitionResult } from "../feature/definition/getDefinition";
 import { getAttributeInterpolationMode } from "../modes/getAttributeInterpolationMode";
 import { getAttributeMode } from "../modes/getAttributeMode";
 import { getAureliaHtmlMode } from "../modes/getAureliaHtmlMode";
