+++ server/src/feature/definition/accessScopeDefinition.ts
@@ -2,16 +2,16 @@ import {
   RepeatForRegionData,
   ViewRegionInfo,
   ViewRegionType,
-} from "./../embeddedLanguages/embeddedSupport";
+} from "../../embeddedLanguages/embeddedSupport";
 /**
  * Defintion[Access Scope]: http://aurelia.io/docs/binding/how-it-works#abstract-syntax-tree
  */
 
 import { TextDocument } from "vscode-languageserver";
-import { aureliaProgram } from "../viewModel/AureliaProgram";
-import { getVirtualDefinition } from "../virtual/virtualDefinition/virtualDefinition";
+import { aureliaProgram } from "../../viewModel/AureliaProgram";
+import { getVirtualDefinition } from "../../virtual/virtualDefinition/virtualDefinition";
 import { DefinitionResult } from "./getDefinition";
-import { Position } from "../embeddedLanguages/languageModes";
+import { Position } from "../../embeddedLanguages/languageModes";
 
 /**
  * Priority
