+++ server/src/modes/getCustomElementMode.ts
@@ -16,7 +16,7 @@ import {
 import { getAureliaVirtualCompletions } from "../virtual/virtualCompletion/virtualCompletion";
 import { getBindablesCompletion } from "../completions/completions";
 import { aureliaProgram } from "../viewModel/AureliaProgram";
-import { DefinitionResult } from "../definition/getDefinition";
+import { DefinitionResult } from "../feature/definition/getDefinition";
 import { connection } from "../server";
 import { camelCase } from "@aurelia/kernel";
 import { getVirtualDefinition } from "../virtual/virtualDefinition/virtualDefinition";
