+++ server/src/server.ts
@@ -53,7 +53,10 @@ import * as path from "path";
 import * as ts from "typescript";
 import { createDiagram } from "./viewModel/createDiagram";
 import { getVirtualDefinition } from "./virtual/virtualDefinition/virtualDefinition";
-import { DefinitionResult, getDefinition } from "./definition/getDefinition";
+import {
+  DefinitionResult,
+  getDefinition,
+} from "./feature/definition/getDefinition";
 import { camelCase } from "@aurelia/kernel";
 
 const globalContainer = new Container();
